+++ macos/SimpleVNKey/SimpleVNKey/EngineBridge/KBEngineWrapper.mm
@@ -12,7 +12,7 @@ static kbengine engine;
 
 @implementation KBEngineWrapper
 - (NSArray*)process: (UInt16)charCode
-                      keycode: (UInt8)keycode
+                      keycode: (UInt16)keycode
                       shiftOrCapPressed: (UInt8)shiftCap
                       otherControlPressed: (bool)otherControl
 {
